// src/lib/interfaces/component.ts
export type ComponentSource = "shadcn" | "MUI" | "Mantine";

export type PreviewType =
  | "card"
  | "button"
  | "modal"
  | "table"
  | "toggle"
  | "progress"
  | "input"
  | "sidebar"
  | "tabs"
  | "badge"
  | "accordion"
  | "skeleton"
  | "tooltip"
  | "dropdown"
  | "alert"
  | "avatar"
  | "pagination"
  | "breadcrumb"
  | "toast"
  | "checkbox";